+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - When enabled, uppon pressing F5, the current localisation will be reloaded and the current window's translation refreshed
 - Added "Report an Issue" button
 - Added info about the translation creator
+- Locale error logs now contain system info
 
 ### Bug Fixes